# Speed up TypeScript perf