<!-- Thanks for submitting a pull request! Please provide enough information so that others can review your pull request. The two fields below are mandatory.

If you're making a change to documentation, do **not** modify a `.html` file directly. Instead, find the corresponding `.pug` file or test case in the `test/docs` directory. -->

**Summary**

Fixes up #10349. Before this PR:

```
$ ./node_modules/.bin/tsc --extendedDiagnostics gh-10349.ts 
Files:                         126
Lines of Library:            27583
Lines of Definitions:        53181
Lines of TypeScript:            20
Lines of JavaScript:             0
Lines of JSON:                   0
Lines of Other:                  0
Nodes of Library:           118533
Nodes of Definitions:       142841
Nodes of TypeScript:            89
Nodes of JavaScript:             0
Nodes of JSON:                   0
Nodes of Other:                  0
Identifiers:                 92346
Symbols:                    177166
Types:                       76289
Instantiations:             334872
Memory used:               204744K
Assignability cache size:    26192
Identity cache size:           175
Subtype cache size:              2
Strict subtype cache size:       4
I/O Read time:               0.01s
Parse time:                  0.55s
ResolveModule time:          0.03s
ResolveTypeReference time:   0.00s
Program time:                0.62s
Bind time:                   0.24s
Check time:                  2.62s
transformTime time:          0.01s
commentTime time:            0.00s
I/O Write time:              0.00s
printTime time:              0.01s
Emit time:                   0.01s
Total time:                  3.50s
```

After this PR:

```
$ ./node_modules/.bin/tsc --extendedDiagnostics gh-10349.ts 
Files:                         126
Lines of Library:            27583
Lines of Definitions:        53167
Lines of TypeScript:            20
Lines of JavaScript:             0
Lines of JSON:                   0
Lines of Other:                  0
Nodes of Library:           118533
Nodes of Definitions:       142530
Nodes of TypeScript:            89
Nodes of JavaScript:             0
Nodes of JSON:                   0
Nodes of Other:                  0
Identifiers:                 92229
Symbols:                    101736
Types:                       47377
Instantiations:              75893
Memory used:               166417K
Assignability cache size:    17401
Identity cache size:           171
Subtype cache size:              2
Strict subtype cache size:       4
I/O Read time:               0.05s
Parse time:                  0.53s
ResolveModule time:          0.03s
ResolveTypeReference time:   0.02s
Program time:                0.66s
Bind time:                   0.23s
Check time:                  1.80s
transformTime time:          0.01s
commentTime time:            0.00s
I/O Write time:              0.00s
printTime time:              0.01s
Emit time:                   0.01s
Total time:                  2.70s
```

We're mostly undoing some more changes from #11650: typing `exists<>`, etc. based on schema is pretty slow. But most of the performance impact, oddly enough, came from adding `extends Query` in `Schema.pre()` and `Schema.post()`. Removing that has brought us back to the perf we had back in 6.3.0 :+1:

**Examples**

<!-- If this code fixes a bug or adds a new feature, provide an example demonstrating the change, unless you added a test. -->