@dawidurbanski 2022-09-05
Just putting it here to let you know this update helped a lot with performance in one very big real project I develop using mongoose. Thanks!